Implemented TerminateCommand for audio
